+++ MigraDoc.RtfRendering/MigraDoc.RtfRendering/RowsRenderer.cs
@@ -62,7 +62,7 @@ namespace MigraDoc.RtfRendering
         if (leftInd == null)
           leftInd = Unit.FromCentimeter(-0.12);
         else
-          leftInd = Unit.FromPoint(-((Unit)leftInd));
+          leftInd = Unit.FromPoint(-(double)((Unit)leftInd));
 
         Cell cell = rows[0].Cells[0];
